Initial commit of the original version used on May 2016 at SACLA
The short notebook introduction presents a very quick overview of the data conversion and analysis that was developed and used during the beamtime in SACLA in May 2016  during the diffraction pump-probe experiment.

Developed by Esposito Vincent (vincent.esposito@psi.ch) and used via Spyder IDE. Thus, certain modifications are not compatible with the initial version by Leo Sala (leonardo.sala@psi.ch), but it's in a TODO list to fix it.